requirements.txt
streamlit
numpy
pandas
Pillow
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git